+++ packages/typespec-go/test/http-specs/payload/pageablegroup/zz_models.go
@@ -10,6 +10,17 @@ type LinkResponse struct {
 	Next *string
 }
 
+type LinkStringResponse struct {
+	// REQUIRED
+	Pets []*Pet
+	Next *string
+}
+
+type ListWithoutContinuationResponse struct {
+	// REQUIRED
+	Pets []*Pet
+}
+
 type NestedLinkResponse struct {
 	// REQUIRED
 	NestedItems *NestedLinkResponseNestedItems