+++ packages/typespec-go/test/http-specs/payload/pageablegroup/zz_models_serde.go
@@ -42,6 +42,64 @@ func (l *LinkResponse) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type LinkStringResponse.
+func (l LinkStringResponse) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "next", l.Next)
+	populate(objectMap, "pets", l.Pets)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type LinkStringResponse.
+func (l *LinkStringResponse) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", l, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "next":
+			err = unpopulate(val, "Next", &l.Next)
+			delete(rawMsg, key)
+		case "pets":
+			err = unpopulate(val, "Pets", &l.Pets)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", l, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type ListWithoutContinuationResponse.
+func (l ListWithoutContinuationResponse) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "pets", l.Pets)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ListWithoutContinuationResponse.
+func (l *ListWithoutContinuationResponse) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", l, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "pets":
+			err = unpopulate(val, "Pets", &l.Pets)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", l, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type NestedLinkResponse.
 func (n NestedLinkResponse) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)