+++ packages/typespec-go/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 0.8.2 (2025-09-10)
+
+### Other Changes
+
+* Updated to the latest tsp toolset.
+
 ## 0.8.1 (2025-09-04)
 
 ### Bugs Fixed